+++ src/main/java/com/example/gccoffee/controller/api/ProductRestController.java
@@ -1,14 +1,15 @@
 package com.example.gccoffee.controller.api;
 
+import com.example.gccoffee.controller.UpdateProductRequest;
 import com.example.gccoffee.model.Category;
 import com.example.gccoffee.model.Product;
 import com.example.gccoffee.service.ProductService;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.UUID;
 
 @RestController
 public class ProductRestController {
@@ -24,6 +25,22 @@ public class ProductRestController {
     return category
       .map(productService::getProductsByCategory)
       .orElse(productService.getAllProducts());
+  }
+
+  //상품 업데이트
+  @PutMapping("/{productId}")
+  public ResponseEntity<Product> updateProduct(@PathVariable UUID productId, @RequestBody UpdateProductRequest request) {
+    Product updatedProduct = productService.updateProduct(productId, request);
+
+    return ResponseEntity.ok(updatedProduct);
+  }
+
+  //상품 삭제
+  @DeleteMapping("/{productId}")
+  public ResponseEntity<Void> deleteProduct(@PathVariable UUID productId) {
+    productService.deleteProduct(productId);
+
+    return ResponseEntity.noContent().build();
   }
 
 }
